Move DropColumn to top level
It's used in multiple places, not all related to the analyzer
