+++ server/src/main/java/io/crate/analyze/AnalyzedAlterTableDropColumn.java
@@ -21,18 +21,11 @@
 
 package io.crate.analyze;
 
-import java.io.IOException;
 import java.util.List;
 import java.util.function.Consumer;
 
-import org.elasticsearch.common.io.stream.StreamInput;
-import org.elasticsearch.common.io.stream.StreamOutput;
-
 import io.crate.expression.symbol.Symbol;
-import io.crate.expression.symbol.Symbols;
-import io.crate.metadata.Reference;
 import io.crate.metadata.doc.DocTableInfo;
-import org.elasticsearch.common.io.stream.Writeable;
 
 public record AnalyzedAlterTableDropColumn(DocTableInfo table, List<DropColumn> columns) implements DDLStatement {
 
@@ -43,18 +36,8 @@ public record AnalyzedAlterTableDropColumn(DocTableInfo table, List<DropColumn> 
 
     @Override
     public void visitSymbols(Consumer<? super Symbol> consumer) {
-    }
-
-    public record DropColumn(Reference ref, boolean ifExists) implements Writeable {
-
-        public DropColumn(StreamInput in) throws IOException {
-            this((Reference) Symbols.fromStream(in), in.readBoolean());
-        }
-
-        @Override
-        public void writeTo(StreamOutput out) throws IOException {
-            Symbols.toStream(ref, out);
-            out.writeBoolean(ifExists);
+        for (DropColumn column : columns) {
+            consumer.accept(column.ref());
         }
     }
 }
